requirements.txt
streamlit
python-dotenv
groq
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
